+++ tests/testthat/test-tl_set_score.R
@@ -106,6 +106,4 @@ test_that("tl_modify works", {
 
   }
 
-
-
 })