+++ NEWS.md
@@ -3,5 +3,5 @@
 * Enable combining treelabel vectors where one is build around a numeric score matrix and the other
 is based on a logical score matrix.
 * Rename `tl_modify_tree` to `tl_tree_modify`
-* Add `tl_tree_filter`
+* Add `tl_tree_filter` and `tl_tree_cut` for modifying the underlying tree
 * Make `tl_eval` more flexible about the output.